move MultiReader to lib
