+++ lib/src/io.rs
@@ -1,7 +1,9 @@
 mod limited_reader;
+mod multi_reader;
 mod multi_writer;
 mod section_reader;
 
 pub use limited_reader::LimitedReader;
+pub use multi_reader::MultiReader;
 pub use multi_writer::MultiWriter;
 pub use section_reader::SectionReader;
